libaio: Improve "cancel" usability.
Kernel may (and does, in recent versions) ignore the "event" argument,
and return EINPROGRESS. In which case user needs to call getEvent to get
cancellation completion, as for all completions. But if it returns the
event, then getEvent may block (if there are no other pending events).
It seems simpler to expose the distinction by return value, as EINPROGRESS
signals a different type of success, but a success nevertheless. To avoid
caller catching all OSErrors without checking errno and taking the wrong
decision, check it internally.