libaio: Fix io_callback_t definition.
iocb parameter is indeed declared to be passed by address and not by value
in the C header:
  typedef void (*io_callback_t)(..., struct iocb *iocb, ...);
This fixes inability to import libaio in some versions of python3.
  https://bugs.python.org/issue39295
  https://bugs.python.org/issue16576
  https://bugs.python.org/issue16575
io_callback_t is not even used here, as it is only involved as parameter
to python functions and gets cast as a void pointer, so no other changes
are expected.